Adicionando encode para senha do usuario
