--- src/main/java/br/com/g1bet/seguranca/UserDetailsServiceImpl.java
@@ -1,34 +1,0 @@
-package br.com.g1bet.seguranca;
-
-import java.util.Optional;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
-
-import br.com.g1bet.model.Usuario;
-import br.com.g1bet.repository.UsuarioRepository;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-	
-	@Autowired
-	private UsuarioRepository usuarioRepository;
-
-	@Override
-	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		
-		Optional<Usuario> usuario = usuarioRepository.findByNome(username);
-		
-		if(usuario.isPresent()) 
-					return new UserDetailsImpl(usuario.get());
-				else 
-					throw new ResponseStatusException(HttpStatus.FORBIDDEN);
-
-	}
-
-}
